貝のギミック実装＆フォルダをギミックフォルダへ移動＆テスト用BGM、SE削除
