+++ Assets/Gimic/shel/Awa.cs.meta
