+++ Assets/Gimic/shel/Shell.cs.meta
