+++ Assets/Gimic/shel/Shell2.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: d590b085effde4f46ae9ffab1260b203
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
