+++ Assets/Gimic/shel/shell.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 367e4aef2f78ce6479fcc771f8590f4d
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
