+++ Assets/Gimic/shel/shell_close.anim
@@ -1,0 +1,148 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!74 &7400000
+AnimationClip:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_Name: shell_close
+  serializedVersion: 6
+  m_Legacy: 0
+  m_Compressed: 0
+  m_UseHighQualityCurve: 1
+  m_RotationCurves: []
+  m_CompressedRotationCurves: []
+  m_EulerCurves: []
+  m_PositionCurves: []
+  m_ScaleCurves: []
+  m_FloatCurves: []
+  m_PPtrCurves:
+  - curve:
+    - time: 0
+      value: {fileID: -6871013029070732337, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    - time: 0.083333336
+      value: {fileID: 1212503278492030864, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    - time: 0.16666667
+      value: {fileID: 8245703108996107866, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    - time: 0.25
+      value: {fileID: -7563952449626683060, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    - time: 0.33333334
+      value: {fileID: 6664546233351484589, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    - time: 0.41666666
+      value: {fileID: 4101124057320942214, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    - time: 0.5
+      value: {fileID: 1915918258706274661, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    - time: 0.5833333
+      value: {fileID: 5670085195202280261, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    - time: 0.6666667
+      value: {fileID: -8661134592498663284, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    - time: 0.75
+      value: {fileID: 5848102774806291847, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    - time: 0.8333333
+      value: {fileID: -1499477454366950912, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    - time: 0.9166667
+      value: {fileID: -2391456164395293452, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    - time: 1
+      value: {fileID: -1505275344485956403, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    - time: 1.0833334
+      value: {fileID: 3742663424880718496, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    - time: 1.1666666
+      value: {fileID: -5316237595889317168, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    - time: 1.25
+      value: {fileID: 7428652858132893235, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    - time: 1.3333334
+      value: {fileID: 3458963824298986957, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    - time: 1.4166666
+      value: {fileID: 106770192944973833, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    - time: 1.5
+      value: {fileID: -8315162655425960620, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    - time: 1.5833334
+      value: {fileID: -8422779670011853078, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    - time: 1.6666666
+      value: {fileID: 7120364321371400250, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    - time: 1.75
+      value: {fileID: -4877517882469302814, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    - time: 1.8333334
+      value: {fileID: -3342969423941313768, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    attribute: m_Sprite
+    path: 
+    classID: 212
+    script: {fileID: 0}
+  m_SampleRate: 12
+  m_WrapMode: 0
+  m_Bounds:
+    m_Center: {x: 0, y: 0, z: 0}
+    m_Extent: {x: 0, y: 0, z: 0}
+  m_ClipBindingConstant:
+    genericBindings:
+    - serializedVersion: 2
+      path: 0
+      attribute: 0
+      script: {fileID: 0}
+      typeID: 212
+      customType: 23
+      isPPtrCurve: 1
+    pptrCurveMapping:
+    - {fileID: -6871013029070732337, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    - {fileID: 1212503278492030864, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    - {fileID: 8245703108996107866, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    - {fileID: -7563952449626683060, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    - {fileID: 6664546233351484589, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    - {fileID: 4101124057320942214, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    - {fileID: 1915918258706274661, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    - {fileID: 5670085195202280261, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    - {fileID: -8661134592498663284, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    - {fileID: 5848102774806291847, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    - {fileID: -1499477454366950912, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    - {fileID: -2391456164395293452, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    - {fileID: -1505275344485956403, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    - {fileID: 3742663424880718496, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    - {fileID: -5316237595889317168, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    - {fileID: 7428652858132893235, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    - {fileID: 3458963824298986957, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    - {fileID: 106770192944973833, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    - {fileID: -8315162655425960620, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    - {fileID: -8422779670011853078, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    - {fileID: 7120364321371400250, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    - {fileID: -4877517882469302814, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+    - {fileID: -3342969423941313768, guid: 1232ee89c717dce48b11841b279ad713, type: 3}
+  m_AnimationClipSettings:
+    serializedVersion: 2
+    m_AdditiveReferencePoseClip: {fileID: 0}
+    m_AdditiveReferencePoseTime: 0
+    m_StartTime: 0
+    m_StopTime: 1.9166667
+    m_OrientationOffsetY: 0
+    m_Level: 0
+    m_CycleOffset: 0
+    m_HasAdditiveReferencePose: 0
+    m_LoopTime: 0
+    m_LoopBlend: 0
+    m_LoopBlendOrientation: 0
+    m_LoopBlendPositionY: 0
+    m_LoopBlendPositionXZ: 0
+    m_KeepOriginalOrientation: 0
+    m_KeepOriginalPositionY: 1
+    m_KeepOriginalPositionXZ: 0
+    m_HeightFromFeet: 0
+    m_Mirror: 0
+  m_EditorCurves: []
+  m_EulerEditorCurves: []
+  m_HasGenericRootTransform: 0
+  m_HasMotionFloatCurves: 0
+  m_Events:
+  - time: 0
+    functionName: StartAnimation
+    data: 
+    objectReferenceParameter: {fileID: 4524685909471883791, guid: d590b085effde4f46ae9ffab1260b203, type: 3}
+    floatParameter: 0
+    intParameter: 0
+    messageOptions: 0
+  - time: 1.8333334
+    functionName: EndClose
+    data: 
+    objectReferenceParameter: {fileID: 4524685909471883791, guid: d590b085effde4f46ae9ffab1260b203, type: 3}
+    floatParameter: 0
+    intParameter: 0
+    messageOptions: 0
