+++ Assets/Gimic/shel/shell_close.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4704eacb67a123c48ad26ea4619dc443
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
