+++ Assets/Gimic/shel/shell_idle.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 88a80e64f7ed64c4ebf42c1e1e743767
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
