+++ Assets/Gimic/shel/shell_open.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: add4f89a78efa154c8e7d43818d31381
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
